+++ CHANGELOG.md
@@ -18,6 +18,10 @@ All notable changes to this project will be documented in this file, in reverse 
 
 ### Fixed
 
+- [#39](https://github.com/phly/http/pull/39) Always attempt to return a Host
+  header from a request if a URI is composed containing a host. This will be
+  true for client-side requests only; server-side will only represent what was
+  present in the incoming request.
 - [#48](https://github.com/phly/http/pull/48) Cache generated URI string after
   first generation.
 - [#42](https://github.com/phly/http/pull/42) Updated `Stream::getSize()` to